% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize_names}
\alias{capitalize_names}
\title{Correctly capitalizes most last names.}
\usage{
capitalize_names(x)
}
\arguments{
\item{x}{A character vector of names}
}
\value{
A character vector of correctly capitalized names
}
\description{
Correctly capitalizes most last names.
}
