% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_upload.R
\name{raw_upload}
\alias{raw_upload}
\title{Raw data to s3}
\usage{
raw_upload(data, election_code, county_code)
}
\arguments{
\item{data}{a data.frame, list, or file path. Data.frames will be uploaded as csvs,
lists will be uploaded as JSON, and files provided as paths will be uploaded as-is}

\item{election_code}{Election code}

\item{county_code}{County code. Typically 5-digit fips}
}
\value{
Nothing
}
\description{
This uploads Raw data to S3, for uploading. Data can then be accessed using \code{\link{raw_download}}
}
