% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_connect.R
\name{results_connect}
\alias{results_connect}
\title{Connects to results database}
\usage{
results_connect(dbname = "precinct_results",
  host = getOption("results_host", NA), port = getOption("results_port",
  NA), username = getOption("results_user", NA),
  password = getOption("results_pass", NA))
}
\arguments{
\item{dbname}{database name}

\item{host}{postgresql host. defaults to \code{getOption("results_host")}}

\item{port}{port \code{getOption("results_port")}}

\item{username}{username \code{getOption("results_user")}}

\item{password}{password \code{getOption("results_pass")}}
}
\description{
Connects to results database
}
