% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_get.R
\name{results_get}
\alias{results_get}
\title{Retrieves election information from external API}
\usage{
results_get(elec_code, state = NULL, county = NULL, cd = NULL)
}
\arguments{
\item{elec_code}{Election code, as a string}

\item{state}{State as either a zero-padded fips code ("04") or a postal code ("AZ")}

\item{county}{County as full 5-digit fips code ("01001")}

\item{cd}{CD as either fips + district number ("0408") or postal code + district number ("AZ08")}
}
\description{
Retrieves election information from external API
}
