% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_upload.R
\name{results_upload}
\alias{results_upload}
\title{Upload results to postgres database}
\usage{
results_upload(df, db, target_table)
}
\arguments{
\item{df}{results dataframe in schema format}

\item{db}{src_postgres to results database}

\item{target_table}{target table in results db to append to}
}
\description{
Upload results to postgres database
}
