% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{schema_precincts}
\alias{schema_precincts}
\title{Returns a character vector of expected precinct names}
\usage{
schema_precincts(county_code)
}
\arguments{
\item{county_code}{5-digit FIPS county code}
}
\value{
A character vector of expected precinct names
}
\description{
Returns a character vector of expected precinct names
}
