% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_fips.R
\name{state_fips}
\alias{state_fips}
\title{Conversion to/from FIPS codes.}
\usage{
state_fips(index)
}
\arguments{
\item{index}{a numeric vector (for conversion from FIPS to postal code) or a character vector (for conversion from postal code to FIPS)}
}
\value{
a numeric or character vector
}
\description{
Converts US states, US territories and Canadian provinces from numeric FIPS code to 2-letter postal code, and vise versa.
}
