% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_schema.R
\name{upload_schema}
\alias{upload_schema}
\title{Checks schema of a results dataframe. Good for creating scrapers}
\usage{
upload_schema(df)
}
\arguments{
\item{df}{a dataframe, in schema format.}
}
\description{
Checks schema of a results dataframe. Good for creating scrapers
}
