% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert.R
\name{upsert}
\alias{upsert}
\title{Upsert a DF into a table}
\usage{
upsert(df, db, target_table)
}
\arguments{
\item{df}{data.frame to upsert}

\item{db}{database src at which the target table resides (can be a \code{tbl} object)}

\item{target_table}{name of the target table (as a character vector)}
}
\description{
Upsert a DF into a table
}
