% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ygresults.R
\docType{package}
\name{ygresults}
\alias{ygresults}
\alias{ygresults-package}
\title{ygresults: rapid results download and tidy}
\description{
ygresults allows for rapid results download and analysis from the yougov precinct data api
}
