% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_connect.R
\name{results_connect}
\alias{results_connect}
\title{Connect to precinct results database}
\usage{
results_connect(dbname, host = getOption("results_host", NA),
  port = getOption("results_port", NA), user = getOption("results_user",
  NA), pass = getOption("results_pass", NA))
}
\arguments{
\item{dbname}{database name}

\item{host}{hostname}

\item{port}{port}

\item{user}{username}

\item{pass}{password}
}
\value{
a database connection object
}
\description{
\code{host, port, user} and \code{pass} can be passed in as \code{options("results_*")}
}
