% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_schema.R
\name{results_schema}
\alias{results_schema}
\title{Checks schema of a results dataframe. Good for creating scrapers}
\usage{
results_schema(df, county_code, is_primary = FALSE)
}
\arguments{
\item{df}{a dataframe, in schema format.}

\item{county_code}{character. 5-digit county FIPS}

\item{is_primary}{logical. is this a primary election?}
}
\description{
Checks schema of a results dataframe. Good for creating scrapers
}
