% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_upload.R
\name{results_upload}
\alias{results_upload}
\title{Uploads results to API}
\usage{
results_upload(df, election_code, county_code, is_primary = FALSE)
}
\arguments{
\item{df}{data.frame with schema columns (see vignette for details)}

\item{election_code}{string election code}

\item{county_code}{string county code (5-digit FIPS)}

\item{is_primary}{deactivates party -- candidate checking for primaries}
}
\description{
Uploads results to API
}
